PropertiesManager
Paths are now stored within a properties file, on boot there will be a check to make sure the properties file already exists on the workstation or not, in case it would not be the case it will make a copy from the default one and place it in the correct position.

Folders are being added in case they are missing.