SaveManager BUG
Fixed an issue that would cause the "recover world" function within the backup system to crash. This was caused by an internal piece of code that would delete the current folder in within the saves section to make sure the backup you place back is clean without the old data.

There is an extra check in place to see if there has to be something removed or not.