Remove world feature modified
Instead of a button that said "remove world" or "delete backup", there is a cross on the right of each save-tile to be used in case the user wants to delete the world from within the app.

There is also a safety feature in place to make sure you can't delete on accident.