Rimworld save manipulator extended
Safety check to make sure that you only max grow planted crops and not the entire map, The entire growth of the whole map would mean you'd have to process 18.000+ entities.

The x.contains("...") has to be extended by each crop that is supposed to pass by.

It could be possible to use the "<sowing>" tag to verify that it was planted. It has to be fact-checked because multiple entities can be sowed, even IF they were not planted by the colonist.