SettingsFrame.java removed and added elsewhere
SettingsFrame.java is removed, but the functionality persists at a different position. The new AddGameFrame "new settings" frame is now responsible for this functionality.

Known issue: The "-" (to remove a game) or "+" (to add a game) does not change upon adding or removing a game from the sidebar list. The purpose is still there, the text just doesn't change.