Frame.java modification + code refactor
Change in Frame.addGame: Instead of removing all the tabs and rebuilding from the ground up, we just create the new cell requested by the user and rewrite the order string to the properties file.

Change in Frame.removeGame: Following Intellij guidelines it was suspicious to remove items from a list within a for loop, we could solve this because we'd only want to remove one entity, the change included to remove the getTabs().remove(i); function with getting the Tab object using getTabs().get(i); and remove the tab outside of the loop using getTabs().remove(COLLECTED_TAB);

Other changes within the class are slight code refactors.