Only show important save files
This new feature will only show requested save formats from a game. This will leave the "total size" feature to show a bigger size compare to what the user will see in the application.

Meaning that there has to be a new feature in place that clusters the save files, thus the user sees ALL files within the save folder.